src/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pocket"
version = "0.1.0"
description = "Unified CLI to materialise Terraform and Ansible configs from a single platform.yaml"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "jsonschema>=4.21",
    "jinja2>=3.1",
    "pydantic>=2.6",
]

[project.scripts]
pocket = "pocket.cli:main"

[project.optional-dependencies]
dev = ["pytest>=8.1"]

[tool.setuptools.packages.find]
where = ["."]
include = ["pocket*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
